requirements.txt
fastapi
uvicorn
groq
gunicorn
opencv-python-headless
requests
pillow
qrcode
pytesseract